changed css class names to fit our group's style guide
