+++ client/src/index.jsx
@@ -4,4 +4,4 @@ import { App } from './app.jsx';
 
 var restaurantId = 'ChIJP5PrLYSAhYARBcWhJXs55P4';
 
-ReactDOM.render(<App restaurantId={restaurantId} restaurant={null}/>, document.getElementById('app'));
+ReactDOM.render(<App restaurantId={restaurantId} restaurant={null}/>, document.getElementById('sidebar-app'));
